config: add option to indicate dns service presence
Adds the config option to set if ipv6 dns service is availiable on the
interface. In some cases the dns service may not be listening on the
ipv6 address of the interface, and thus should not be announced to clients.

Signed-off-by: Paizhuo Chen <cc@icpz.dev>
(cherry picked from commit 694148a87fa5a0c964286fdaac1fa490a2fb530b)