dhcpv6-ia: fix prefix delegation behavior
When an IPv6 address change is triggered each PD assignment is checked
if it is still consistent with the updated IPv6 prefix list.
If not consistent anymore a reconfigure is triggered for the assignment
and a best effort is made to assign a new IA_PD prefix.
If it not possible anymore to assign an IA_PD prefix delete the PD
assignment now so it will result into a NO BINDING status code for the
given IA_PD in the DHCPv6 reply when the client tries to renew the IA_PD prefix.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>
(cherry picked from commit 9d5e37912820443492f87f07fb68edeed206c044)